feat: change long to integer

+++ src/main/java/com/reviews/restaurant/dto/ReviewResponseDTO.java
@@ -26,7 +26,7 @@ public class ReviewResponseDTO {
     private Double totalScore;
     private String comments;
     private RestaurantResponseDTO restaurant;
-    private Long idUser;
+    private Integer idUser;
     private List<ImageResponseDTO> images;
 
 }
